Move v1beta1 CRD implementation into v1beta1 package. Also do some minor code cleanup
