+++ client/src/main/scala/skuber/api/client/impl/KubernetesClientImpl.scala
@@ -22,8 +22,8 @@ import skuber.json.format.apiobj.statusReads
 import skuber.json.format.{apiVersionsFormatReads, deleteOptionsFmt, namespaceListFmt}
 
 import javax.net.ssl.SSLContext
-import skuber.apiextensions.CustomResourceDefinition.Scope
 import skuber.config.SkuberConfig
+import skuber.apiextensions.v1beta1.CustomResourceDefinition.Scope
 
 import scala.concurrent.duration._
 import scala.concurrent.{ExecutionContext, Future, Promise}
