--- client/src/main/scala/skuber/apiextensions/package.scala
@@ -1,8 +1,0 @@
-package skuber
-
-/**
-  * @author David O'Riordan
-  */
-package object apiextensions {
-  type CustomResourceDefinitionList=ListResource[CustomResourceDefinition]
-}
