+++ examples/src/main/scala/skuber/examples/customresources/CreateCRD.scala
@@ -2,7 +2,7 @@ package skuber.examples.customresources
 
 import skuber.{K8SException, k8sInit}
 import skuber.ResourceSpecification.Scope
-import skuber.apiextensions.CustomResourceDefinition
+import skuber.apiextensions.v1beta1.CustomResourceDefinition
 import akka.actor.ActorSystem
 import scala.concurrent.ExecutionContextExecutor
 import scala.util.{Failure, Success}
